// libs/common/src/database/postgres-error.enum.ts
export enum PostgresErrorCode {
  UniqueViolation = '23505',
  ForeignKeyViolation = '23503',
  NotNullViolation = '23502',
  CheckViolation = '23514',
  InvalidTextRepresentation = '22P02',
  NumericValueOutOfRange = '22003',
  DivisionByZero = '22012',
  DataException = '22000',
  IntegrityConstraintViolation = '23000',
  SyntaxErrorOrAccessRuleViolation = '42000',
  InsufficientResources = '53000',
  DiskFull = '53100',
  OutOfMemory = '53200',
  TooManyConnections = '53300',
  ConfigurationLimitExceeded = '53400',
  OperatorIntervention = '57000',
  SystemError = '58000',
  IoError = '58030',
  UndefinedTable = '42P01',
  UndefinedColumn = '42703',
}